+++ src/sora-cpp-sdk/include/sora/dyn/dyn.h
@@ -1,7 +1,7 @@
 #ifndef DYN_DYN_H_
 #define DYN_DYN_H_
 
-#include <iostream>
+#include <iostream>  // IWYU pragma: export
 #include <map>
 #include <memory>
 #include <string>