+++ src/sora-cpp-sdk/include/sora/hwenc_nvcodec/nvcodec_decoder_cuda.h
@@ -1,11 +1,12 @@
 #ifndef SORA_HWENC_NVCODEC_NVCODEC_DECODER_CUDA_H_
 #define SORA_HWENC_NVCODEC_NVCODEC_DECODER_CUDA_H_
 
+#include <cstdint>
 #include <memory>
+#include <string>
 
 // clang-format off
 #include "sora/fix_cuda_noinline_macro_error.h"
-#include <string>
 // clang-format on
 
 #include "sora/cuda_context.h"