+++ src/sora-cpp-sdk/src/open_h264_video_encoder.cpp
@@ -14,9 +14,14 @@
 
 #include "sora/open_h264_video_encoder.h"
 
+#include <algorithm>
+#include <cstddef>
+#include <cstdint>
+#include <cstring>
 #include <limits>
 #include <memory>
 #include <string>
+#include <utility>
 #include <vector>
 
 #if defined(_WIN32)
@@ -29,30 +34,43 @@
 
 // WebRTC
 #include <absl/container/inlined_vector.h>
-#include <absl/strings/match.h>
+#include <absl/memory/memory.h>
 #include <absl/types/optional.h>
+#include <api/environment/environment.h>
 #include <api/environment/environment_factory.h>
-#include <api/transport/rtp/dependency_descriptor.h>
+#include <api/scoped_refptr.h>
+#include <api/units/data_rate.h>
+#include <api/video/encoded_image.h>
 #include <api/video/i420_buffer.h>
+#include <api/video/video_bitrate_allocation.h>
+#include <api/video/video_bitrate_allocator.h>
 #include <api/video/video_codec_constants.h>
+#include <api/video/video_codec_type.h>
+#include <api/video/video_frame.h>
+#include <api/video/video_frame_buffer.h>
+#include <api/video/video_frame_type.h>
 #include <api/video_codecs/scalability_mode.h>
+#include <api/video_codecs/sdp_video_format.h>
+#include <api/video_codecs/video_codec.h>
 #include <api/video_codecs/video_encoder.h>
 #include <common_video/h264/h264_bitstream_parser.h>
 #include <common_video/libyuv/include/webrtc_libyuv.h>
+#include <media/base/media_constants.h>
 #include <modules/video_coding/codecs/h264/include/h264.h>
+#include <modules/video_coding/codecs/h264/include/h264_globals.h>
+#include <modules/video_coding/codecs/interface/common_constants.h>
 #include <modules/video_coding/include/video_codec_interface.h>
 #include <modules/video_coding/include/video_error_codes.h>
 #include <modules/video_coding/svc/create_scalability_structure.h>
 #include <modules/video_coding/svc/scalable_video_controller.h>
-#include <modules/video_coding/utility/quality_scaler.h>
 #include <modules/video_coding/utility/simulcast_rate_allocator.h>
 #include <modules/video_coding/utility/simulcast_utility.h>
 #include <rtc_base/checks.h>
 #include <rtc_base/logging.h>
-#include <rtc_base/time_utils.h>
 #include <system_wrappers/include/metrics.h>
-#include <third_party/libyuv/include/libyuv/convert.h>
-#include <third_party/libyuv/include/libyuv/scale.h>
+
+// libyuv
+#include <libyuv/scale.h>
 
 // OpenH264
 #include <wels/codec_api.h>
@@ -932,7 +950,7 @@ std::unique_ptr<webrtc::VideoEncoder> CreateOpenH264VideoEncoder(
     }
   }
 
-  return std::make_unique<webrtc::OpenH264VideoEncoder>(
+  return absl::make_unique<webrtc::OpenH264VideoEncoder>(
       webrtc::CreateEnvironment(), settings, std::move(openh264));
 }