+++ src/sora-cpp-sdk/src/scalable_track_source.cpp
@@ -10,17 +10,21 @@
 
 #include "sora/scalable_track_source.h"
 
-#include <algorithm>
+#include <cstdint>
+#include <optional>
 
 // WebRTC
+#include <api/media_stream_interface.h>
 #include <api/scoped_refptr.h>
 #include <api/video/i420_buffer.h>
+#include <api/video/video_frame.h>
 #include <api/video/video_frame_buffer.h>
 #include <api/video/video_rotation.h>
-#include <rtc_base/logging.h>
+#include <media/base/adapted_video_track_source.h>
+#include <rtc_base/time_utils.h>
 
 // libyuv
-#include <libyuv.h>
+#include <libyuv/rotate.h>
 
 namespace sora {